0.1 penambahan laporan ++

+++ app/Http/Controllers/BeasiswaController.php
@@ -4,6 +4,8 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 
+use Illuminate\Support\Facades\File;
+
 use App\Beasiswa;
 use App\Pendaftaran;
 
@@ -14,7 +16,7 @@ class BeasiswaController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
-    public function index()
+    public function index() 
     {
         $data = Beasiswa::orderBy('id', 'DESC')->get();
 
@@ -49,6 +51,12 @@ class BeasiswaController extends Controller
             'jenis' => $request->jenis,
             'kontrak_beasiswa' => $request->kontrak_beasiswa,
         ]; 
+
+        if ($file = $request->file('persyaratan')) {
+            $nama_file = 'Persayaratan_'. time() . $file->getClientOriginalName();
+            $file->move(public_path() . '/File/Beasiswa/Persyaratan/', $nama_file);  
+            $data['persyaratan'] = $nama_file;
+        }
 
         Beasiswa::create($data);
         
@@ -101,6 +109,17 @@ class BeasiswaController extends Controller
             'kontrak_beasiswa' => $request->kontrak_beasiswa,
         ];
 
+        if ($file = $request->persyaratan){
+
+            if ($beasiswa->persyaratan) {
+                File::delete('File/Beasiswa/Persyaratan/' . $beasiswa->persyaratan);
+            }
+
+            $nama_file = 'Persyaratan_' . time() . $file->getClientOriginalName();
+            $file->move(public_path() . '/File/Beasiswa/Persyaratan/', $nama_file);  
+            $data['persyaratan'] = $nama_file;
+        }
+
         $beasiswa->update($data);
 
         return redirect()->back()->with('success', 'Berhasil Merubah Data');
@@ -116,6 +135,11 @@ class BeasiswaController extends Controller
     {
         $beasiswa = Beasiswa::find($id);
 
+        $pendaftaran = Pendaftaran::where('id_beasiswa', $id)->first();
+        if($pendaftaran){
+
+            return redirect()->back()->with('error', 'Tidak dapat menghapus data karena sudah ada mahasiswa yang mendaftar beasiswa ini');
+        }
         $beasiswa->delete();
 
         return redirect()->back()->with('success', 'Berhasil Menghapus Data');
